+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `extract_namespaces` function in `util.py`
 ### Added
 - `DMetaBaseError` added to `dmeta/__init__.py`
+- `overwrite_metadata` function added to `functions.py`
 ### Changed
+- `update` function in `functions.py` refactored
+- `clear` function in `functions.py` refactored
 - `README.md` updated
 - GitHub actions are limited to the `dev` and `main` branches
 - `Python 3.13` added to `test.yml`